Move mesh state into separate file
